+++ src/main/scala/scuff/ddd/Repository.scala
@@ -16,9 +16,9 @@ trait Repository[AR <: AggregateRoot] {
    * @param revision Load a specific revision, or None for most current
    * @return The requested revision or [[scuff.ddd.UnknownIdException]]
    */
-  def load(id: AR#ID, revision: Option[Long] = None): Future[AR]
-  final def load(id: (AR#ID, Long)): Future[AR] = load(id._1, Some(id._2))
-  final def load(id: AR#ID, revision: Long): Future[AR] = load(id, Some(revision))
+  def load(id: AR#ID, revision: Option[Int] = None): Future[AR]
+  final def load(id: (AR#ID, Int)): Future[AR] = load(id._1, Some(id._2))
+  final def load(id: AR#ID, revision: Int): Future[AR] = load(id, Some(revision))
 
   /**
    * Update aggregate. Changes are committed and any revision conflict is
@@ -29,7 +29,7 @@ trait Repository[AR <: AggregateRoot] {
    * @param basedOnRevision Revision, which update will be based on
    * @param updateBlock The transaction code block. This may be executed multiple times if concurrent updates occur
    */
-  def update[T](id: AR#ID, basedOnRevision: Long = Long.MaxValue)(updateBlock: AR ⇒ T)(implicit metadata: Map[String, String] = Map.empty): Future[(T, Long)]
+  def update[T](id: AR#ID, basedOnRevision: Int = Int.MaxValue)(updateBlock: AR ⇒ T)(implicit metadata: Map[String, String] = Map.empty): Future[(T, Int)]
 
   /**
    * Insert new aggregate root and publish committed events.
