+++ src/main/scala/scuff/eventual/DuplicateRevisionException.scala
@@ -4,4 +4,4 @@ package scuff.eventual
  * There has been a concurrent revision update between
  * loading and updating the stream.
  */
-class DuplicateRevisionException(id: Any, revision: Long) extends RuntimeException("Revision %d already exists: %s".format(revision, id))
+class DuplicateRevisionException(id: Any, revision: Int) extends RuntimeException("Revision %d already exists: %s".format(revision, id))
