+++ src/main/scala/scuff/eventual/ddd/FixedIntervalSnapshots.scala
@@ -15,7 +15,7 @@ trait FixedIntervalSnapshots[ID, AR <: AggregateRoot, CAT] extends EventStoreRep
 
   protected override def assumeSnapshotCurrent = revisionInterval <= 2
 
-  protected abstract override def saveSnapshot(id: AR#ID, revision: Long, state: S) {
+  protected abstract override def saveSnapshot(id: AR#ID, revision: Int, state: S) {
     if ((revision + 1) % revisionInterval == 0) {
       super.saveSnapshot(id, revision, state)
     }
